+++ src/deluge/model/song/song.cpp
@@ -173,6 +173,7 @@ Song::Song() : backedUpParamManagers(sizeof(BackedUpParamManager)) {
 	// Setup reverb temp variables
 	reverbRoomSize = (float)30 / 50;
 	reverbDamp = (float)36 / 50;
+	reverbHPF = 0;
 	reverbLPF = (float)50 / 50;
 	reverbWidth = 1;
 	reverbPan = 0;
@@ -1142,16 +1143,19 @@ weAreInArrangementEditorOrInClipInstance:
 	uint32_t roomSize = AudioEngine::reverb.getRoomSize() * (uint32_t)2147483648u;
 	uint32_t damping = AudioEngine::reverb.getDamping() * (uint32_t)2147483648u;
 	uint32_t width = AudioEngine::reverb.getWidth() * (uint32_t)2147483648u;
+	uint32_t hpf = AudioEngine::reverb.getHPF() * (uint32_t)2147483648u;
 	uint32_t lpf = AudioEngine::reverb.getLPF() * (uint32_t)2147483648u;
 
 	roomSize = std::min(roomSize, (uint32_t)2147483647);
 	damping = std::min(damping, (uint32_t)2147483647);
 	width = std::min(width, (uint32_t)2147483647);
+	hpf = std::min(hpf, (uint32_t)2147483647);
 	lpf = std::min(lpf, (uint32_t)2147483647);
 
 	writer.writeAttribute("roomSize", roomSize);
 	writer.writeAttribute("dampening", damping);
 	writer.writeAttribute("width", width);
+	writer.writeAttribute("hpf", hpf);
 	writer.writeAttribute("lpf", lpf);
 	writer.writeAttribute("pan", AudioEngine::reverbPan);
 	writer.writeAttribute("model", util::to_underlying(model));
@@ -1363,6 +1367,14 @@ Error Song::readFromFile(Deserializer& reader) {
 						}
 						reverbWidth = (float)widthInt / 2147483648u;
 						reader.exitTag("width");
+					}
+					else if (!strcmp(tagName, "hpf")) {
+						reverbHPF = (float)reader.readTagOrAttributeValueInt() / 2147483648u;
+						reader.exitTag("hpf");
+					}
+					else if (!strcmp(tagName, "lpf")) {
+						reverbLPF = (float)reader.readTagOrAttributeValueInt() / 2147483648u;
+						reader.exitTag("lpf");
 					}
 					else if (!strcmp(tagName, "pan")) {
 						reverbPan = reader.readTagOrAttributeValueInt();
@@ -3295,6 +3307,27 @@ void Song::replaceInstrument(Instrument* oldOutput, Instrument* newOutput, bool 
 	if (newOutput->type != OutputType::KIT && oldOutput->type != OutputType::KIT) {
 		((MelodicInstrument*)newOutput)->midiInput = ((MelodicInstrument*)oldOutput)->midiInput;
 		((MelodicInstrument*)oldOutput)->midiInput.clear();
+	}
+
+	// if we're replacing the same output type
+	if (newOutput->type == oldOutput->type) {
+		// Migrate any midi learned params for synth and kit clip preset changes
+		// For kit clips it will migrate only kit affect entire midi learned params
+		// Scenario:
+		// - you create a clip
+		// - you midi learn a controller to that clip's params
+		// - you then go to change the preset for that clip
+		// - you expect that you can continue controlling the same params for the new preset
+		Sound* oldSound = (Sound*)oldOutput->toModControllable();
+		if (oldSound) {
+			int32_t numKnobs = oldSound->midiKnobArray.getNumElements();
+			if (numKnobs) {
+				Sound* newSound = (Sound*)newOutput->toModControllable();
+				newSound->midiKnobArray.cloneFrom(&oldSound->midiKnobArray);
+				oldSound->midiKnobArray.deleteAtIndex(0, numKnobs);
+				oldSound->ensureInaccessibleParamPresetValuesWithoutKnobsAreZero(this);
+			}
+		}
 	}
 
 	Output* outputRecordingOldOutput = oldOutput->getOutputRecordingThis();