+++ src/deluge/model/song/song.h
@@ -397,6 +397,7 @@ public:
 	// Reverb params to be stored here between loading and song being made the active one
 	dsp::Reverb::Model model;
 	float reverbRoomSize;
+	float reverbHPF;
 	float reverbLPF;
 	float reverbDamp;
 	float reverbWidth;